no more up flags
